backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.26
botocore==1.37.26
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.2
django-admin-logs==1.4.0
django-celery==3.1.17
django-filter==25.1
django-multiselectfield==0.1.13
django-rangefilter==0.9.1
django-ses==4.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx==0.2.4
Faker==37.1.0
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
kombu==5.5.2
lxml==5.3.2
msgpack==1.1.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
